apply ts on BrandFilterResponseDTO
