--- src/data/response_dto/brand/BrandFilterResponseDTO.js
@@ -1,19 +1,0 @@
-import BrandResponseDTO from './BrandResponseDTO';
-
-class BrandFilterResponseDTO {
-    constructor({ firstInitial, brands }) {
-        this.firstInitial = firstInitial;
-        this.brands = brands;
-    }
-
-    static create(brandFilterDTO) {
-        return new BrandFilterResponseDTO({
-            firstInitial: brandFilterDTO.firstInitial,
-            brands: brandFilterDTO.brands.map(
-                (it) => new BrandResponseDTO(it.brandIdx, it.name)
-            ),
-        });
-    }
-}
-
-module.exports = BrandFilterResponseDTO;
